refactor: Encapsulate common TV availability checker logic
